+++ src/App/Fossa/Config/Analyze.hs
@@ -6,6 +6,8 @@ module App.Fossa.Config.Analyze (
   BinaryDiscovery (..),
   ExperimentalAnalyzeConfig (..),
   ForceVendoredDependencyRescans (..),
+  ForceFirstPartyScans (..),
+  ForceNoFirstPartyScans (..),
   IATAssertion (..),
   DynamicLinkInspect (..),
   IncludeAll (..),
@@ -54,6 +56,7 @@ import App.Fossa.Subcommand (EffStack, GetCommonOpts (getCommonOpts), GetSeverit
 import App.Fossa.VSI.Types qualified as VSI
 import App.Types (
   BaseDir,
+  FirstPartyScansFlag (..),
   OverrideDynamicAnalysisBinary (..),
   OverrideProject (OverrideProject),
   ProjectMetadata (projectLabel),
@@ -107,6 +110,8 @@ import Types (ArchiveUploadType (..), LicenseScanPathFilters (..), TargetFilter)
 -- CLI flags, for use with 'Data.Flag'
 data DeprecatedAllowNativeLicenseScan = DeprecatedAllowNativeLicenseScan deriving (Generic)
 data ForceVendoredDependencyRescans = ForceVendoredDependencyRescans deriving (Generic)
+data ForceFirstPartyScans = ForceFirstPartyScans deriving (Generic)
+data ForceNoFirstPartyScans = ForceNoFirstPartyScans deriving (Generic)
 
 data BinaryDiscovery = BinaryDiscovery deriving (Generic)
 data IncludeAll = IncludeAll deriving (Generic)
@@ -187,6 +192,8 @@ data AnalyzeCliOpts = AnalyzeCliOpts
   , analyzeSkipVSIGraphResolution :: [VSI.Locator]
   , analyzeBaseDir :: FilePath
   , analyzeDynamicGoAnalysisType :: GoDynamicTactic
+  , analyzeForceFirstPartyScans :: Flag ForceFirstPartyScans
+  , analyzeForceNoFirstPartyScans :: Flag ForceNoFirstPartyScans
   }
   deriving (Eq, Ord, Show)
 
@@ -213,6 +220,7 @@ data AnalyzeConfig = AnalyzeConfig
   , includeAllDeps :: Flag IncludeAll
   , noDiscoveryExclusion :: Flag NoDiscoveryExclusion
   , overrideDynamicAnalysis :: OverrideDynamicAnalysisBinary
+  , firstPartyScansFlag :: FirstPartyScansFlag
   }
   deriving (Eq, Ord, Show, Generic)
 
@@ -260,6 +268,8 @@ cliParser =
     <*> many skipVSIGraphResolutionOpt
     <*> baseDirArg
     <*> experimentalUseV3GoResolver
+    <*> flagOpt ForceFirstPartyScans (long "experimental-force-first-party-scans" <> help "Force first party scans")
+    <*> flagOpt ForceNoFirstPartyScans (long "experimental-block-first-party-scans" <> help "Block first party scans. This can be used to forcibly turn off first-party scans if your organization defaults to first-party scans.")
 
 data GoDynamicTactic
   = GoModulesBasedTactic
@@ -379,6 +389,12 @@ mergeStandardOpts maybeConfig envvars cliOpts@AnalyzeCliOpts{..} = do
       experimentalCfgs = collectExperimental maybeConfig cliOpts
       vendoredDepsOptions = collectVendoredDeps maybeConfig cliOpts
       dynamicAnalysisOverrides = OverrideDynamicAnalysisBinary $ envCmdOverrides envvars
+  firstPartyScansFlag <-
+    case (fromFlag ForceFirstPartyScans analyzeForceFirstPartyScans, fromFlag ForceNoFirstPartyScans analyzeForceNoFirstPartyScans) of
+      (True, True) -> fatalText "You provided both the --experimental-force-first-party-scans and --experimental-block-first-party-scans flags. Only one of these flags may be used"
+      (True, _) -> pure FirstPartyScansOnFromFlag
+      (_, True) -> pure FirstPartyScansOffFromFlag
+      (False, False) -> pure FirstPartyScansUseDefault
 
   AnalyzeConfig
     <$> basedir
@@ -394,6 +410,7 @@ mergeStandardOpts maybeConfig envvars cliOpts@AnalyzeCliOpts{..} = do
     <*> pure analyzeIncludeAllDeps
     <*> pure analyzeNoDiscoveryExclusion
     <*> pure dynamicAnalysisOverrides
+    <*> pure firstPartyScansFlag
 
 collectFilters ::
   ( Has Diagnostics sig m