+++ src/Control/Carrier/FossaApiClient/Internal/LicenseScanning.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE RecordWildCards #-}
 
 module Control.Carrier.FossaApiClient.Internal.LicenseScanning (
+  getSignedFirstPartyScanUrl,
   getSignedLicenseScanUrl,
   finalizeLicenseScan,
   uploadLicenseScanResult,
+  uploadFirstPartyScanResult,
 ) where
 
 import Control.Algebra (Has)
@@ -13,8 +15,20 @@ import Control.Effect.FossaApiClient (PackageRevision (..))
 import Control.Effect.Lift (Lift)
 import Control.Effect.Reader (Reader, ask)
 import Control.Monad (void)
+import Data.List.NonEmpty qualified as NE
 import Fossa.API.Types (ApiOpts, ArchiveComponents, SignedURL)
-import Srclib.Types (LicenseSourceUnit)
+import Srclib.Types (FullSourceUnit, LicenseSourceUnit)
+
+getSignedFirstPartyScanUrl ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has (Reader ApiOpts) sig m
+  ) =>
+  PackageRevision ->
+  m SignedURL
+getSignedFirstPartyScanUrl PackageRevision{..} = do
+  apiOpts <- ask
+  API.getSignedFirstPartyScanURL apiOpts packageVersion packageName
 
 getSignedLicenseScanUrl ::
   ( Has (Lift IO) sig m
@@ -47,3 +61,13 @@ uploadLicenseScanResult ::
   m ()
 uploadLicenseScanResult signedUrl licenseSourceUnit = do
   void $ API.licenseScanResultUpload signedUrl licenseSourceUnit
+
+uploadFirstPartyScanResult ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  ) =>
+  SignedURL ->
+  NE.NonEmpty FullSourceUnit ->
+  m ()
+uploadFirstPartyScanResult signedUrl fullSourceUnits = do
+  void $ API.firstPartyScanResultUpload signedUrl fullSourceUnits